Revert "Merge branch 'main' into userAuthenticationUIComponents"
This reverts commit d6a1af826cf94becd2263cbec9c294779346b4f7, reversing
changes made to b8a4a64e8cc739ab9f8a8fc1c5a2e27f9248ac09.
