--- .github/workflows/dev-metrics.yml
@@ -1,71 +1,0 @@
-name: Daily Project Metrics Generation
-
-on:
-  schedule:
-    - cron: "0 0 * * *" # Runs at 00:00 UTC every day
-  workflow_dispatch: # Allows manual trigger from the GitHub UI
-
-jobs:
-  generate_and_commit_metrics:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write # Give the default GITHUB_TOKEN write permission to commit and push the added or changed files to the repository.
-      pull-requests: write # Requires you go to Settings>Actions>Workflow Permissions>"Allow GitHub Actions to create and approve pull requests"
-    steps:
-      - name: Checkout Local Repo
-        uses: actions/checkout@v4
-      - name: Checkout INSO GH Metrics code
-        uses: actions/checkout@v4
-        with:
-          repository: Poggecci/INSO-GH-API-QUERY
-          path: inso-gh-query-metrics
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: "3.11"
-      - name: Install poetry
-        uses: abatilo/actions-poetry@v2
-      - name: Setup a local virtual environment (if no poetry.toml file)
-        run: |
-          poetry config virtualenvs.create true --local
-          poetry config virtualenvs.in-project true --local
-      - uses: actions/cache@v3
-        name: Define a cache for the virtual environment based on the dependencies lock file
-        with:
-          path: ./.venv
-          key: venv-${{ hashFiles('poetry.lock') }}
-      - name: Install the project dependencies
-        run: |
-          cd inso-gh-query-metrics
-          poetry install
-        shell: bash
-      - name: Run metrics generator
-        run: |
-          cd inso-gh-query-metrics
-          poetry run python src/generateMilestoneMetricsForActions.py ${{ github.workspace }}/gh_metrics_config.json
-        shell: bash
-        env:
-          GITHUB_API_TOKEN: ${{ secrets.GH_API_TOKEN }}
-          ORGANIZATION: ${{ github.repository_owner }}
-      - name: Move Generated Metrics to Source Repo
-        run: |
-          mkdir -p ./metrics
-          mv inso-gh-query-metrics/*${{ github.repository_owner }}.md metrics/
-        shell: bash
-      - name: Auto Commit Results
-        uses: stefanzweifel/git-auto-commit-action@v4
-        with:
-          commit_message: "[Automated] Milestone Metrics Update"
-          branch: inso-metrics
-          push_options: "--force"
-          create_branch: true
-          file_pattern: metrics/*.md
-          commit_user_name: GitHub Action
-          commit_user_email: action@github.com
-          commit_author: GitHub Action <action@github.com>
-      - name: Create Pull Request
-        run: |
-          gh pr create --base ${{ github.ref }} --head "inso-metrics" --title "[Automated] Milestone Metrics Update" --body "This is an automated pull request to update the milestone metrics." --repo ${{ github.repository }} || echo "Command failed: Unable to create pull request. It might be due to the branch already existing or lack of permissions."
-        env:
-          GH_TOKEN: ${{ github.token }}
-        shell: bash
