+++ safe-rum/app/(tabs)/explore.tsx
