--- safe-rum/app/index.tsx
@@ -1,8 +1,0 @@
-import { Redirect } from "expo-router";
-
-// App starts here
-const StartPage = () => {
-  return <Redirect href="/home" />;
-};
-
-export default StartPage;
